pyproject.toml
[tool.poetry]
name = "apisample"
version = "0.1.0"
description = ""
authors = ["kustiktmkirill <happykust@list.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.0"}
drf-yasg = "^1.21.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
